Develop

+++ modules/document/document.controller.php
@@ -234,6 +234,7 @@ class documentController extends document
 		if(!$output->toBool()) return $output;
 		// Register it if no given document_srl exists
 		if(!$obj->document_srl) $obj->document_srl = getNextSequence();
+		elseif(!$manual_inserted && !$isRestore && !checkUserSequence($obj->document_srl)) return new Object(-1, 'msg_not_permitted');
 
 		$oDocumentModel = getModel('document');
 		// Set to 0 if the category_srl doesn't exist
@@ -267,7 +268,7 @@ class documentController extends document
 		}
 		// If the tile is empty, extract string from the contents.
 		settype($obj->title, "string");
-		if($obj->title == '') $obj->title = cut_str(strip_tags($obj->content),20,'...');
+		if($obj->title == '') $obj->title = cut_str(trim(strip_tags(nl2br($obj->content))),20,'...');
 		// If no tile extracted from the contents, leave it untitled.
 		if($obj->title == '') $obj->title = 'Untitled';
 		// Remove XE's own tags from the contents.
@@ -332,7 +333,10 @@ class documentController extends document
 		$oDB->commit();
 
 		// return
-		$this->addGrant($obj->document_srl);
+		if(!$manual_inserted)
+		{
+			$this->addGrant($obj->document_srl);
+		}
 		$output->add('document_srl',$obj->document_srl);
 		$output->add('category_srl',$obj->category_srl);
 
@@ -398,7 +402,15 @@ class documentController extends document
 		if(!$obj->commentStatus) $obj->commentStatus = 'DENY';
 		if($obj->commentStatus == 'DENY') $this->_checkCommentStatusForOldVersion($obj);
 		if($obj->allow_trackback!='Y') $obj->allow_trackback = 'N';
-		if($obj->homepage &&  !preg_match('/^[a-z]+:\/\//i',$obj->homepage)) $obj->homepage = 'http://'.$obj->homepage;
+		if($obj->homepage)
+		{
+			$obj->homepage = removeHackTag($obj->homepage);
+			if(!preg_match('/^[a-z]+:\/\//i',$obj->homepage))
+			{
+				$obj->homepage = 'http://'.$obj->homepage;
+			}
+		}
+		
 		if($obj->notify_message != 'Y') $obj->notify_message = 'N';
 		
 		// can modify regdate only manager
@@ -503,30 +515,33 @@ class documentController extends document
 			return $output;
 		}
 		// Remove all extra variables
-		$this->deleteDocumentExtraVars($source_obj->get('module_srl'), $obj->document_srl, null, Context::getLangType());
-		// Insert extra variables if the document successfully inserted.
-		$extra_keys = $oDocumentModel->getExtraKeys($obj->module_srl);
-		if(count($extra_keys))
+		if(Context::get('act')!='procFileDelete')
 		{
-			foreach($extra_keys as $idx => $extra_item)
+			$this->deleteDocumentExtraVars($source_obj->get('module_srl'), $obj->document_srl, null, Context::getLangType());
+			// Insert extra variables if the document successfully inserted.
+			$extra_keys = $oDocumentModel->getExtraKeys($obj->module_srl);
+			if(count($extra_keys))
 			{
-				$value = '';
-				if(isset($obj->{'extra_vars'.$idx}))
+				foreach($extra_keys as $idx => $extra_item)
 				{
-					$tmp = $obj->{'extra_vars'.$idx};
-					if(is_array($tmp))
-						$value = implode('|@|', $tmp);
-					else
-						$value = trim($tmp);
+					$value = NULL;
+					if(isset($obj->{'extra_vars'.$idx}))
+					{
+						$tmp = $obj->{'extra_vars'.$idx};
+						if(is_array($tmp))
+							$value = implode('|@|', $tmp);
+						else
+							$value = trim($tmp);
+					}
+					else if(isset($obj->{$extra_item->name})) $value = trim($obj->{$extra_item->name});
+					if($value == NULL) continue;
+					$this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, $idx, $value, $extra_item->eid);
 				}
-				else if(isset($obj->{$extra_item->name})) $value = trim($obj->{$extra_item->name});
-				if(!isset($value)) continue;
-				$this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, $idx, $value, $extra_item->eid);
 			}
+			// Inert extra vars for multi-language support of title and contents.
+			if($extra_content->title) $this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, -1, $extra_content->title, 'title_'.Context::getLangType());
+			if($extra_content->content) $this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, -2, $extra_content->content, 'content_'.Context::getLangType());
 		}
-		// Inert extra vars for multi-language support of title and contents.
-		if($extra_content->title) $this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, -1, $extra_content->title, 'title_'.Context::getLangType());
-		if($extra_content->content) $this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, -2, $extra_content->content, 'content_'.Context::getLangType());
 		// Update the category if the category_srl exists.
 		if($source_obj->get('category_srl') != $obj->category_srl || $source_obj->get('module_srl') == $logged_info->member_srl)
 		{
@@ -555,7 +570,7 @@ class documentController extends document
 		if($oCacheHandler->isSupport())
 		{
 			//remove document item from cache
-			$cache_key = 'document_item:'.$obj->document_srl;
+			$cache_key = 'document_item:'. getNumberingPath($obj->document_srl) . $obj->document_srl;
 			$oCacheHandler->delete($cache_key);
 		}
 
@@ -646,7 +661,7 @@ class documentController extends document
 		$oCacheHandler = CacheHandler::getInstance('object');
 		if($oCacheHandler->isSupport())
 		{
-			$cache_key = 'document_item:'.$document_srl;
+			$cache_key = 'document_item:'. getNumberingPath($document_srl) . $document_srl;
 			$oCacheHandler->delete($cache_key);
 		}
 
@@ -793,6 +808,8 @@ class documentController extends document
 		$oCacheHandler = CacheHandler::getInstance('object');
 		if($oCacheHandler->isSupport())
 		{
+			$cache_key = 'document_item:'. getNumberingPath($oDocument->document_srl) . $oDocument->document_srl;
+			$oCacheHandler->delete($cache_key);
 		}
 
 		return $output;
@@ -805,12 +822,16 @@ class documentController extends document
 	 */
 	function updateReadedCount(&$oDocument)
 	{
+		// Pass if Crawler access
+		if(isCrawler()) return false;
+		
 		$document_srl = $oDocument->document_srl;
 		$member_srl = $oDocument->get('member_srl');
 		$logged_info = Context::get('logged_info');
-		// Call a trigger when the read count is updated (after)
-		$output = ModuleHandler::triggerCall('document.updateReadedCount', 'after', $oDocument);
-		if(!$output->toBool()) return $output;
+
+		// Call a trigger when the read count is updated (before)
+		$trigger_output = ModuleHandler::triggerCall('document.updateReadedCount', 'before', $oDocument);
+		if(!$trigger_output->toBool()) return $trigger_output;
 
 		// Pass if read count is increaded on the session information
 		if($_SESSION['readed_document'][$document_srl]) return false;
@@ -827,10 +848,32 @@ class documentController extends document
 			$_SESSION['readed_document'][$document_srl] = true;
 			return false;
 		}
+
+		$oDB = DB::getInstance();
+		$oDB->begin();
+
 		// Update read counts
 		$args = new stdClass;
 		$args->document_srl = $document_srl;
 		$output = executeQuery('document.updateReadedCount', $args);
+
+		// Call a trigger when the read count is updated (after)
+		$outptrigger_outputut = ModuleHandler::triggerCall('document.updateReadedCount', 'after', $oDocument);
+		if(!$trigger_output->toBool())
+		{
+			$oDB->rollback();
+			return $trigger_output;
+		}
+
+		$oDB->commit();
+
+		$oCacheHandler = CacheHandler::getInstance('object');
+		if($oCacheHandler->isSupport())
+		{
+			//remove document item from cache
+			$cache_key = 'document_item:'. getNumberingPath($document_srl) . $document_srl;
+			$oCacheHandler->delete($cache_key);
+		}
 
 		// Register session
 		$_SESSION['readed_document'][$document_srl] = true;
@@ -1097,18 +1140,31 @@ class documentController extends document
 
 		$oDB->commit();
 
+		$oCacheHandler = CacheHandler::getInstance('object');
+		if($oCacheHandler->isSupport())
+		{
+			//remove document item from cache
+			$cache_key = 'document_item:'. getNumberingPath($document_srl) . $document_srl;
+			$oCacheHandler->delete($cache_key);
+		}
+
 		// Leave in the session information
 		$_SESSION['voted_document'][$document_srl] = true;
 
 		// Return result
+		$output = new Object();
 		if($point > 0)
 		{
-			return new Object(0, 'success_voted');
+			$output->setMessage('success_voted');
+			$output->add('voted_count', $obj->after_point);
 		}
 		else
 		{
-			return new Object(0, 'success_blamed');
+			$output->setMessage('success_blamed');
+			$output->add('blamed_count', $obj->after_point);
 		}
+		
+		return $output;
 	}
 
 	/**
@@ -1238,6 +1294,14 @@ class documentController extends document
 		{
 			$args->update_order = -1*getNextSequence();
 			$args->last_updater = $last_updater;
+
+			$oCacheHandler = CacheHandler::getInstance('object');
+			if($oCacheHandler->isSupport())
+			{
+				//remove document item from cache
+				$cache_key = 'document_item:'. getNumberingPath($document_srl) . $document_srl;
+				$oCacheHandler->delete($cache_key);
+			}
 		}
 
 		return executeQuery('document.updateCommentCount', $args);
@@ -1254,6 +1318,14 @@ class documentController extends document
 		$args = new stdClass;
 		$args->document_srl = $document_srl;
 		$args->trackback_count = $trackback_count;
+
+		$oCacheHandler = CacheHandler::getInstance('object');
+		if($oCacheHandler->isSupport())
+		{
+			//remove document item from cache
+			$cache_key = 'document_item:'. getNumberingPath($document_srl) . $document_srl;
+			$oCacheHandler->delete($cache_key);
+		}
 
 		return executeQuery('document.updateTrackbackCount', $args);
 	}
@@ -1358,6 +1430,30 @@ class documentController extends document
 		if(!$output->toBool()) return $output;
 
 		$this->makeCategoryFile($category_info->module_srl);
+		// remvove cache
+		$oCacheHandler = CacheHandler::getInstance('object');
+		if($oCacheHandler->isSupport())
+		{
+			$page = 0;
+			while(true) {
+				$args = new stdClass();
+				$args->category_srl = $category_srl;
+				$args->list_count = 100;
+				$args->page = ++$page;
+				$output = executeQuery('document.getDocumentList', $args, array('document_srl'));
+
+				if($output->data == array())
+					break;
+
+				foreach($output->data as $val)
+				{
+					//remove document item from cache
+					$cache_key = 'document_item:'. getNumberingPath($val->document_srl) . $val->document_srl;
+					$oCacheHandler->delete($cache_key);
+				}
+			}
+		}
+
 		// Update category_srl of the documents in the same category to 0
 		$args = new stdClass();
 		$args->target_category_srl = 0;
@@ -1464,6 +1560,7 @@ class documentController extends document
 		$this->updateCategory($cur_args);
 		// Category information
 		$next_args = new stdClass;
+		$next_args->category_srl = $next_category->category_srl;
 		$next_args->list_order = $list_order;
 		$next_args->title = $next_category->title;
 		$this->updateCategory($next_args);
@@ -2151,7 +2248,7 @@ class documentController extends document
 
 				if($type=='move') $purl = sprintf("<a href=\"%s\" onclick=\"window.open(this.href);return false;\">%s</a>", $oDocument->getPermanentUrl(), $oDocument->getPermanentUrl());
 				else $purl = "";
-				$content .= sprintf("<div>%s</div><hr />%s<div style=\"font-weight:bold\">%s</div>%s",$message_content, $purl, $oDocument->getTitleText(), $oDocument->getContent(false, false, false));
+				$content = sprintf("<div>%s</div><hr />%s<div style=\"font-weight:bold\">%s</div>%s",$message_content, $purl, $oDocument->getTitleText(), $oDocument->getContent(false, false, false));
 
 				$oCommunicationController->sendMessage($sender_member_srl, $oDocument->get('member_srl'), $title, $content, false);
 			}
